+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v2.5.7
+1. fix: 通过 [urlParams](https://wproxy.org/whistle/rules/urlParams.html) 和 [pathReplace](https://wproxy.org/whistle/rules/pathReplace.html) 修改请求URL参数的问题
+2. style: Network 右键菜单添加 `Actions>Mark` 标记抓包数据
+3. refactor: `statusCode` 移入 `rule` 里面跟 `file` 等协议同级
+
+# v2.5.6
+1. style: Plugins 添加 `ReinstallAll` 按钮，可以copy插件安装命令
+2. fix: 还原匹配顺序，修复：https://github.com/avwo/whistle/issues/421
+
 # v2.5.5
 1. fix: `Node >= 14.1` 无法使用http2问题
 2. fix: 某些规则可能会被插件返回的规则覆盖问题